Move permissions to the top level
